added logic file, and added logic to get positions, and init and took out logic from main routes
