+++ app/controllers/main_routes/main_routes.py
@@ -85,31 +85,6 @@ def departmentPortal(org=None,account=None):
                            positions = positionsList,
                            posURL = posURL)
 
-@main_bp.route('/department/<org>/<account>/positions', methods=['GET'])
-def managePositions(org, account):
-    try:
-        dept = Department.get(Department.ORG == org, Department.ACCOUNT == account)
-    except DoesNotExist:
-        return render_template('errors/404.html'), 404
-    
-    if not g.currentUser.isLaborAdmin:
-        if not SupervisorDepartment.select().where(
-            (SupervisorDepartment.supervisor == g.currentUser.supervisor) &
-            (SupervisorDepartment.department == dept.departmentID)
-        ).exists():
-            return render_template('errors/403.html'), 403
-
-    positions = (PositionHistory.select()
-                                .where((PositionHistory.department == dept) &
-                                       (PositionHistory.status == "Active"))
-                                .order_by(PositionHistory.positionTitle.asc()))
-
-    return render_template('main/managePositions.html',
-                           department = dept,
-                           department_name = dept.DEPT_NAME,
-                           positions = positions
-                           )
-
 @main_bp.route('/supervisorPortal/download', methods=['POST'])
 def downloadSupervisorPortalResults():
     '''
